requirements.txt
ollama>=0.1.8
pytrends>=4.9.2
APScheduler>=3.10.4
Flask>=3.0.3
Jinja2>=3.1.4
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-slugify>=8.0.4
moviepy>=1.0.3
gTTS>=2.5.1
Pillow>=10.4.0
numpy>=1.26.0
google-api-python-client>=2.139.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
pandas>=2.2.2
